fix: PR review fixes
